pyproject.toml
[project]
name = "skill-agent"
version = "0.1.0"
description = "Skill Builder Demo — agentic pipeline from request to publishable skill"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.13.0",
    "httpx>=0.28.1",
    "pydantic>=2.7.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
